+++ components/MobileMenu/index.tsx
@@ -1,0 +1,1 @@
+export { MobileMenu } from './MobileMenu'
