+++ components/MobileMenu/mobilemenu-readme.md
@@ -1,4 +1,4 @@
-** ReadMe DOC for using the DNA MainMenu Component **
+** ReadMe DOC for using the DNA MobileMenu Component **
 
 ** dependencies in use **
 eve
@@ -11,7 +11,7 @@ classnames
 pageWrapId - the id of the element wrapping the rest of content on your page
 (except elements with fixed positioning),placed after the component
 example:
-<MainMenu pageWrapId="page-wrap" />
+<MobileMenu pageWrapId="page-wrap" />
 <main id="page-wrap">
 .
 .
@@ -21,7 +21,7 @@ example:
 2. outerContainerId - the id of the element containing everything,including the menu component
 example:
 <div id="outer-container">
-<MainMenu pageWrapId="page-wrap" outerContainerId="outer-container" />
+<MobileMenu pageWrapId="page-wrap" outerContainerId="outer-container" />
 <main id="page-wrap">
 .
 .
@@ -34,13 +34,13 @@ To have it open from the right, use the right prop.
 It's just a boolean so you don't need to specify a value. 
 Then set the position of the button using CSS.
 example:
-<MainMenu right />
+<MobileMenu right />
 
 4. width
 You can specify the width of the menu with the width prop. The default is 300.
-<MainMenu width={200} />
-<MainMenu width={'280px'} />
-<MainMenu width={‘20%’} />
+<MobileMenu width={200} />
+<MobileMenu width={'280px'} />
+<MobileMenu width={‘20%’} />
 
 5. isOpen
 You can control whether the sidebar is open or closed with the isOpen prop. 
@@ -48,7 +48,7 @@ This is useful if you need to close the menu after a user clicks on an item in i
 for example, or if you want to open the menu from some other button in addition to the standard burger icon.
 The default value is false.
 example:
-<MainMenu isOpen={true} />
+<MobileMenu isOpen={true} />
 
 6. onOpen
 If you keep the menu state yourself it might be convenient
@@ -57,7 +57,7 @@ something that should open the menu.
 Called when:
  the user clicks on the burger icon
 example:
-<MainMenu onOpen={handleOnOpen} />
+<MobileMenu onOpen={handleOnOpen} />
 Note: The menu will NOT open automatically 
 if you pass this prop,
 so you must handle it yourself.
@@ -71,7 +71,7 @@ Called when:
 The user clicks on the cross icon
 The user clicks on the overlay
 The user hits the escape key
-<MainMenu onClose={ handleOnClose } />
+<MobileMenu onClose={ handleOnClose } />
 Note: The menu will NOT close automatically 
 if you pass this prop, 
 so you must handle it yourself.
@@ -86,7 +86,7 @@ var isMenuOpen = function(state) {
   return state.isOpen;
 };
 
-<MainMenu onStateChange={ isMenuOpen } />
+<MobileMenu onStateChange={ isMenuOpen } />
 
 9. disableCloseOnEsc
 By default, the menu will close when the Escape key 
@@ -97,7 +97,7 @@ for example if you're implementing a responsive
 menu that behaves differently depending on the 
 browser width.
 
-<MainMenu disableCloseOnEsc />
+<MobileMenu disableCloseOnEsc />
 
 10. customOnKeyDown
 For more control over global keypress functionality,
@@ -116,7 +116,7 @@ this.setState({areMenusOpen: false});
 }
 };
 
-<MainMenu customOnKeyDown={closeAllMenusOnEsc} isOpen={areMenusOpen} />
+<MobileMenu customOnKeyDown={closeAllMenusOnEsc} isOpen={areMenusOpen} />
 Note: Using this prop will disable all the default 
 'close on Escape' functionality, so you will need 
 to handle this (including determining which key was 
@@ -124,16 +124,16 @@ pressed) yourself.
 
 11. noOverlay
 You can turn off the default overlay with noOverlay.
-<MainMenu noOverlay />
+<MobileMenu noOverlay />
 You can disable the overlay click event (i.e. prevent overlay clicks from closing the menu) with disableOverlayClick.
 This can either be a boolean, or a function that returns a boolean.
-<MainMenu disableOverlayClick />
-<MainMenu disableOverlayClick={() => shouldDisableOverlayClick() } />
+<MobileMenu disableOverlayClick />
+<MobileMenu disableOverlayClick={() => shouldDisableOverlayClick() } />
 
 12. noTransition
 You can disable all transitions/animations by passing 
 noTransition.
-<MainMenu noTransition />
+<MobileMenu noTransition />
 This is useful if you want the menu to remain open 
 across re-mounts, for example during SPA route changes.
 
@@ -143,8 +143,8 @@ burger and cross icons with custom ReactElements.
 Pass them as the customBurgerIcon and customCrossIcon 
 props respectively.
 example:
-<MainMenu customBurgerIcon={ <img src="img/icon.svg" /> } />
-<MainMenu customCrossIcon={ <img src="img/cross.svg" /> } />
+<MobileMenu customBurgerIcon={ <img src="img/icon.svg" /> } />
+<MobileMenu customCrossIcon={ <img src="img/cross.svg" /> } />
 This can be useful if you want exclusive external control 
 of the menu, using the isOpen prop.
 14. id and className
@@ -154,20 +154,20 @@ to the rendered menu's outermost element.
 This is not required for any functionality, 
 but could be useful for things like styling 
 with CSS modules.
-<MainMenu id={ "sidebar" } className={ "my-menu" } />
+<MobileMenu id={ "sidebar" } className={ "my-menu" } />
 You can also pass custom 
 classNames to the other elements:
-<MainMenu burgerButtonClassName={ "my-class" } />
-<MainMenu burgerBarClassName={ "my-class" } />
-<MainMenu crossButtonClassName={ "my-class" } />
-<MainMenu crossClassName={ "my-class" } />
-<MainMenu menuClassName={ "my-class" } />
-<MainMenu morphShapeClassName={ "my-class" } />
-<MainMenu itemListClassName={ "my-class" } />
-<MainMenu overlayClassName={ "my-class" } />
+<MobileMenu burgerButtonClassName={ "my-class" } />
+<MobileMenu burgerBarClassName={ "my-class" } />
+<MobileMenu crossButtonClassName={ "my-class" } />
+<MobileMenu crossClassName={ "my-class" } />
+<MobileMenu menuClassName={ "my-class" } />
+<MobileMenu morphShapeClassName={ "my-class" } />
+<MobileMenu itemListClassName={ "my-class" } />
+<MobileMenu overlayClassName={ "my-class" } />
 And to the html and body elements (applied when the menu is open):
-<MainMenu htmlClassName={ "my-class" } />
-<MainMenu bodyClassName={ "my-class" } />
+<MobileMenu htmlClassName={ "my-class" } />
+<MobileMenu bodyClassName={ "my-class" } />
 Note: Passing these props will prevent the menu 
 from applying styles to the html or body elements 
 automatically.
@@ -177,7 +177,7 @@ By default, the menu will set focus on the first
 item when opened. This is to help with keyboard 
 navigation. If you don't want this functionality, 
 you can pass the disableAutoFocus prop.
-<MainMenu disableAutoFocus />
+<MobileMenu disableAutoFocus />
 
 16. itemListElement
 The menu's children are all wrapped in a nav 
@@ -186,7 +186,7 @@ the most common use case for this component.
 However, it's a general purpose sidebar, 
 so you can change this to a div if you're 
 not using it for navigation:
-<MainMenu itemListElement="div" />
+<MobileMenu itemListElement="div" />
 
 17. styles
 All the animations are handled internally 
@@ -309,17 +309,17 @@ var styles = {
   }
 }
 
-<MainMenu styles={ styles } />
+<MobileMenu styles={ styles } />
 
 ** usage **
-import {MainMenu} from '../components'
+import {MobileMenu} from '../components'
 <div id="outter-container">
- <MainMenu pageWrapId="page-wrap" outterContainerId="outter-container" animationType="slide" right={false} styles={styles}>
+ <MobileMenu pageWrapId="page-wrap" outterContainerId="outter-container" animationType="slide" right={false} styles={styles}>
         <a id="home" className="menu-item" href="/">Home</a>
         <a id="about" className="menu-item" href="/about">About</a>
         <a id="contact" className="menu-item" href="/contact">Contact</a>
         <a onClick={ this.showSettings } className="menu-item--small" href="">Settings</a>
-</MainMenu>
+</MobileMenu>
 <main id="page-wrap">
 .
 .
