+++ components/MobileMenu/snapsvgImporter.ts
