+++ components/MobileMenu/types/CrossIcon.d.ts
