+++ package.json
@@ -47,19 +47,20 @@
     "react-dom": "17.0.1",
     "react-query": "^3.6.0",
     "react-query-auth": "^0.3.0",
+    "react-sticky-el": "^2.0.6",
     "resolve-from": "^5.0.0",
     "snapsvg-cjs": "^0.0.6",
     "webpack": "^4.46.0",
-    "react-sticky-el": "^2.0.6",
     "yup": "^0.32.9"
   },
   "devDependencies": {
+    "@babel/plugin-syntax-export-default-from": "^7.14.5",
     "@types/node": "^14.14.25",
     "@types/react": "^17.0.1",
-    "cross-env": "^7.0.3",
     "@types/react-sticky-el": "^1.0.2",
     "@typescript-eslint/eslint-plugin": "^4.25.0",
     "@typescript-eslint/parser": "^4.25.0",
+    "cross-env": "^7.0.3",
     "eslint": "^7.27.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
